% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maei_clu_rand.R
\name{get_maei_clust_rand}
\alias{get_maei_clust_rand}
\title{Title}
\usage{
get_maei_clust_rand(
  vr,
  dist,
  nvoters,
  cluster,
  z,
  s10,
  s01 = NULL,
  pi = NULL,
  exp_ac0 = 1
)
}
\arguments{
\item{vr}{A data frame contains variable district and the number of voters in the district.}

\item{dist}{character variable to specify variable district}

\item{nvoters}{character variable to specify the number of voters}

\item{cluster}{character variable to specify the cluster}

\item{z}{treatment}

\item{s10}{can be number or proportion}

\item{s01}{default NUll: in case one}

\item{pi}{pi}

\item{exp_ac0}{default 1 most conservative}
}
\description{
Description
}
\examples{
\dontrun{
data(rv)  # input data
get_maei_clust_rand (vr = rv, dist = "d", nvoters = "n_voters", cluster = "c", z = Z,  s10=0.2, s01 = 30, pi = 0.3, exp_ac0 = 1)
}


}
\references{
xxx
}
