% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maei_ind_rand.R
\name{get_maei_ind_rand}
\alias{get_maei_ind_rand}
\title{Title}
\usage{
get_maei_ind_rand(vr, dist, nvoters, s10, s01 = NULL, exp_ac0 = 1)
}
\arguments{
\item{vr}{A data frame contains variable district and the number of voters in the district.}

\item{dist}{character to specify district variable}

\item{nvoters}{character to specify the number of voters vairable}

\item{s10}{can be number or proportion}

\item{s01}{default NUll: in case one}

\item{exp_ac0}{default 1 most conservative}
}
\description{
Description
}
\examples{
\dontrun{
data(rv1)  # input data
get_maei_ind_rand(vr = rv1, dist = "d", nvoters = "n_voters", s10 = 20)
}
}
\references{
xxx
}
