% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{rv}
\alias{rv}
\title{Sample dataset}
\format{
A sample data frame
\describe{
  \item{d}{code for each country}
  \item{n_voters}{the preference for income redistribution from 1 to 10}
  \item{c}{individuals' income level}
}
}
\usage{
rv
}
\description{
A dataset containing xxx.
}
\keyword{datasets}
