% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv1.R
\docType{data}
\name{rv1}
\alias{rv1}
\title{Example dataset without clusters}
\format{
A sample data frame
\describe{
  \item{d}{Denotes each district}
  \item{n_voters}{The number of registered voters in each district.}
}
}
\usage{
rv1
}
\description{
A data set with example data used in vignettes for for the function get_maei_ind_rand.
}
\keyword{datasets}
