% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maei_clu_rand.R
\name{get_maei_clust_rand}
\alias{get_maei_clust_rand}
\title{Title}
\usage{
get_maei_clust_rand(
  vr,
  dist,
  nvoters,
  cluster,
  Z,
  s10,
  s01 = NULL,
  pi = NULL,
  exp_ac0 = 1
)
}
\arguments{
\item{vr}{a data frame contains variables: district and corresponding the number of voters in the district.}

\item{dist}{a character to specify the column name of the district variable in the vr.}

\item{nvoters}{a character to specify the column number of the number of voters in the vr.}

\item{cluster}{a character to to specify the column number of the cluster variable in the vr.}

\item{Z}{a vector to denote the treatment assignments.}

\item{s10}{a number or vector to denote the individuals exposed to the treatment because it is assigned experimentally. It can be the exact number or proportion (between 0 and 1).}

\item{s01}{a number or vector to denote theindividuals not exposed to the treatment because is assigned experimentally. It can be the exact number or proportion (between 0 and 1). Default value is NUll which means it is the case one  in which a researcher designs and implements an intervention that would otherwise not have occurred. If it is not NUll, it means case 2 in which some intervention by an NGO or IG is modified to include an experimental component.}

\item{exp_ac0}{a number or vector to denote the expectation of untreated potential outcome. The default value is one which will return the most conservative bound.}

\item{a}{number or vector to to measure researchers’ ex-ante beliefs about the proportion of voters that could respond to treatment (or some manifestation thereof) in clusters where allocation of the intervention is not changed by the experiment.}
}
\description{
Description
}
\examples{
\dontrun{
data(rv)  # input data
require(randomizr)
Z = block_ra(blocks = rv$d, prob = 1/3) ### generate treatment assignments

get_maei_clust_rand (vr = rv, dist = "d",
nvoters = "n_voters",
cluster = "c", Z = Z,
s10=0.2, s01 = 30, pi = 0.3,
 exp_ac0 = 1)

get_maei_clust_rand (vr = rv, dist = "d",
nvoters = "n_voters",
cluster = "c", Z = Z,
s10=round(runif(n = 150,min = 25,max = 75)),
s01 = 30,
pi = 0.1)
}


}
\references{
xxx
}
