% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maei_ind_rand.R
\name{get_maei_ind_rand}
\alias{get_maei_ind_rand}
\title{Title}
\usage{
get_maei_ind_rand(
  vr,
  dist,
  nvoters,
  s10,
  s01 = NULL,
  exp_ac0 = 1,
  psi = NULL,
  ...
)
}
\arguments{
\item{vr}{a data frame contains variables: district and corresponding the number of voters in the district.}

\item{dist}{a character to specify the column name of the district variable in the vr.}

\item{nvoters}{a character to specify the column number of the number of voters in the vr.}

\item{s10}{a number or vector to denote the individuals exposed to the treatment because it is assigned experimentally. It can be the exact number or proportion (between 0 and 1).}

\item{s01}{a number or vector to denote theindividuals not exposed to the treatment because is assigned experimentally. It can be the exact number or proportion (between 0 and 1). Default value is NUll which means it is the case one  in which a researcher designs and implements an intervention that would otherwise not have occurred. If it is not NUll, it means case 2 in which some intervention by an NGO or IG is modified to include an experimental component.}

\item{exp_ac0}{a number or vector to denote the expectation of untreated potential outcome. The default value is one which will return the most conservative bound.}

\item{psi}{a number or vector specifying “margin to pivotality”, as minimum change in vote share, as a proportion of registered voters, at which a different officeholder would be elected in district. If psi > 2MAEI (Maximal Aggregate Electoral Impact), an experiment could not change the ultimate electoral outcome (the output result will show "PASS" the decision rule); in contrast, if psi < 2MAEI, the experiment could affect the ultimate electoral outcome (the output result will show "FAIL").}
}
\description{
Description
}
\examples{
### input data
data(rv1)

### specify s10 as a constant number across districts
get_maei_ind_rand(vr = rv1, dist = "d",
nvoters = "n_voters", s10 = 200)

### let s10 be different proportion and specify s01 = 10 (thus case 2)
set.seed(10)
get_maei_ind_rand(vr = rv1, dist = "d",
nvoters = "n_voters", s10 = runif(10), s01 = 10)

### add margin to pivotality say psi = 0.3
get_maei_ind_rand(vr = rv1, dist = "d",
nvoters = "n_voters", s10 = 0.13, s01 = 10, psi = 0.3)

}
\references{
xxx
}
