% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv.R
\docType{data}
\name{rv}
\alias{rv}
\title{Example dataset with clusters}
\format{
A sample data frame
\describe{
  \item{d}{Denotes each district}
  \item{n_voters}{The number of registered voters in each cluster in each district.}
  \item{c}{Denotes the clusters within each district.}
}
}
\usage{
rv
}
\description{
A data set with example data used in vignettes for for the function get_maei_rand.
}
\keyword{datasets}
